+++ fractals/src/main/java/me/catmousedog/fractals/canvas/Mouse.java
@@ -6,7 +6,12 @@ import java.awt.event.MouseListener;
 import javax.swing.SwingUtilities;
 
 import me.catmousedog.fractals.fractals.LinearTransform;
+import me.catmousedog.fractals.ui.JPInterface;
+import me.catmousedog.fractals.ui.Logger;
 
+/**
+ * the mouse listener for interacting with the canvas
+ */
 public class Mouse implements MouseListener {
 
 	/**
@@ -15,35 +20,48 @@ public class Mouse implements MouseListener {
 	private final Canvas canvas;
 
 	/**
-	 * factor to multiply the zoom with each time the user clicks
+	 * the logger instance
 	 */
-	private double zoomFactor = 2;
+	private final Logger logger;
 
-	public Mouse(Canvas canvas) {
+	/**
+	 * the user interface, used to {@link JPInterface#render()} and retrieve user
+	 * data
+	 */
+	private JPInterface jpi;
+
+	private double zoomFactor = 1.5;
+
+	public Mouse(Canvas canvas, Logger logger) {
 		this.canvas = canvas;
+		this.logger = logger;
 	}
 
 	@Override
-	public void mouseReleased(MouseEvent e) {
-		e = SwingUtilities.convertMouseEvent(e.getComponent(), e, canvas);
+	public void mouseReleased(MouseEvent me) {
+		me = SwingUtilities.convertMouseEvent(me.getComponent(), me, canvas);
 
 		LinearTransform transform = canvas.getTransform();
-		double[] t = transform.apply(e.getX(), e.getY());
-		
-		// lmb zoom in
-		if (e.getButton() == MouseEvent.BUTTON1) {
-			transform.setTranslation(t[0], t[1]);
-		}
-
-		// rmb zoom out
-		if (e.getButton() == MouseEvent.BUTTON3) {
-			transform.setTranslation(t[0], t[1]);
-		}
-
+		double[] t = transform.apply(me.getX(), me.getY());
 		transform.setTranslation(t[0], t[1]);
-		double newZoom = transform.getm() / zoomFactor;
-		transform.setScalar(newZoom, newZoom);
-		canvas.render(null);
+
+		// manually save zoomFactor as it is needed for the transformation
+		try {
+			zoomFactor = Double.parseDouble(jpi.getZoomJTF().getText());
+		} catch (NumberFormatException e) {
+			logger.log("zoom factor must be a valid double");
+		}
+
+		// lmb (zoom in)
+		if (me.getButton() == MouseEvent.BUTTON1)
+			transform.zoom(1 / zoomFactor);
+
+		// rmb (zoom out)
+		if (me.getButton() == MouseEvent.BUTTON3)
+			transform.zoom(zoomFactor);
+
+		jpi.update();
+		jpi.renderNow();
 	}
 
 	@Override
@@ -62,12 +80,7 @@ public class Mouse implements MouseListener {
 	public void mouseExited(MouseEvent e) {
 	}
 
-	/**
-	 * sets the zoomFactor
-	 * 
-	 * @param z the new zoomFactor
-	 */
-	public void setZoomFactor(double z) {
-		zoomFactor = z;
+	public void setJPI(JPInterface jpi) {
+		this.jpi = jpi;
 	}
 }