+++ fractals/src/main/java/me/catmousedog/fractals/main/Settings.java
@@ -1,0 +1,15 @@
+package me.catmousedog.fractals.main;
+
+public class Settings {
+
+	private boolean render_on_changes = true;
+
+	public void setRender_on_changes(boolean render_on_changes) {
+		this.render_on_changes = render_on_changes;
+	}
+
+	public boolean isRender_on_changes() {
+		return render_on_changes;
+	}
+
+}